+++ src/components/Button/view/default/Button.tsx
@@ -1,14 +1,22 @@
 import React, { FC } from 'react';
-import { StyledButton } from './style/Button.style';
+import { StyledButton } from './styles';
 import { ButtonProps } from '../../types';
 
 //TODO: Implement css only animation
 export const Button: FC<ButtonProps> = ({
     htmlType,
     type = 'default',
+    className,
     ...props
 }) => {
-    return <StyledButton atlanticType={type} type={htmlType} {...props} />;
+    return (
+        <StyledButton
+            atlanticType={type}
+            type={htmlType}
+            className={className}
+            {...props}
+        />
+    );
 };
 
 Button.displayName = `Button`;