+++ src/components/Button/view/simple/SimpleButton.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
 import { SimpleButtonProps } from '../../types';
-import { StyledSimpleButton } from './style';
+import { StyledSimpleButton } from './styles';
 
 export const SimpleButton: FC<SimpleButtonProps> = StyledSimpleButton;