+++ src/components/Indicator/view/Indicator.tsx
@@ -1,9 +1,14 @@
 import React, { FC } from 'react';
-import { StyledIndicator } from './style/Indicator.style';
-import { IndicatorProps } from '../types/Indicator.types';
+import { StyledIndicator } from './styles';
+import { IndicatorProps } from '../types';
 
-export const Indicator: FC<IndicatorProps> = ({ type, size, ...props }) => (
-    <StyledIndicator type={type} size={size} {...props} />
+export const Indicator: FC<IndicatorProps> = ({
+    type,
+    size,
+    className,
+    ...props
+}) => (
+    <StyledIndicator type={type} size={size} className={className} {...props} />
 );
 
 Indicator.displayName = `Indicator`;