+++ src/components/TextArea/TextArea.tsx
@@ -1,5 +1,5 @@
 import React, { FC, forwardRef } from 'react';
-import { StyledTextArea } from './style';
+import { StyledTextArea } from './styles';
 import { TextAreaProps } from './types';
 
 export type TextAreaType = FC<TextAreaProps>;
@@ -7,6 +7,6 @@ export type TextAreaType = FC<TextAreaProps>;
 export const TextArea: TextAreaType = forwardRef<
     HTMLTextAreaElement,
     TextAreaProps
->(({ ...props }, ref) => {
-    return <StyledTextArea {...props} ref={ref} />;
+>(({ className, ...props }, ref) => {
+    return <StyledTextArea className={className} {...props} ref={ref} />;
 });